work/py-harness/packages/coding-harness/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "coding-harness"
version = "0.1.0"
description = "Coding-agent scaffolding on top of the pyharness SDK: built-in tools, AGENTS.md, settings, skills, named agents, extensions, CLI. One concrete harness built on the pi-py kernel."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "pi-py contributors" }]
dependencies = [
    "pyharness",
    "pyyaml>=6.0",
    "httpx>=0.27",
    "python-frontmatter>=1.1",
]

[project.optional-dependencies]
extract = ["trafilatura>=1.8"]

[project.scripts]
pyharness = "coding_harness.cli:main"

[tool.setuptools.packages.find]
where = ["src"]
